requirements.txt
argcomplete==1.10.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
compressed-rtf==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
google-search-results==2.4.1
idna==3.3
IMAPClient==2.1.0
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
lxml==4.8.0
MarkupSafe==2.1.1
nltk==3.7
oauthlib==3.2.0
olefile==0.46
pdfminer.six==20191110
pdfminer3==2018.12.3.0
Pillow==9.1.0
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
PyPDF2==1.27.2
python-decouple==3.6
python-pptx==0.6.21
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
SpeechRecognition==3.8.1
sqlparse==0.4.2
textract==1.6.5
tqdm==4.63.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
xlrd==1.2.0
XlsxWriter==3.0.3
